Move the primitive module out of the properties mod into draw
