+++ src/draw/primitive/ellipse.rs
@@ -1,6 +1,7 @@
+use crate::draw::primitive::Primitive;
 use crate::draw::properties::spatial::{dimension, orientation, position};
 use crate::draw::properties::{
-    spatial, ColorScalar, Draw, Drawn, IntoDrawn, LinSrgba, Primitive, SetColor, SetDimensions,
+    spatial, ColorScalar, Draw, Drawn, IntoDrawn, LinSrgba, SetColor, SetDimensions,
     SetOrientation, SetPosition,
 };
 use crate::draw::{self, Drawing};
