+++ src/draw/primitive/mesh.rs
@@ -1,6 +1,7 @@
 use crate::draw::mesh::vertex::IntoVertex;
+use crate::draw::primitive::Primitive;
 use crate::draw::properties::spatial::{self, orientation, position};
-use crate::draw::properties::{Draw, Drawn, IntoDrawn, Primitive, SetOrientation, SetPosition};
+use crate::draw::properties::{Draw, Drawn, IntoDrawn, SetOrientation, SetPosition};
 use crate::draw::{self, Drawing};
 use crate::geom;
 use crate::math::BaseFloat;
