+++ src/draw/primitive/rect.rs
@@ -1,7 +1,8 @@
 use crate::draw;
+use crate::draw::primitive::Primitive;
 use crate::draw::properties::spatial::{dimension, orientation, position};
 use crate::draw::properties::{
-    spatial, ColorScalar, Draw, Drawn, IntoDrawn, LinSrgba, Primitive, SetColor, SetDimensions,
+    spatial, ColorScalar, Draw, Drawn, IntoDrawn, LinSrgba, SetColor, SetDimensions,
     SetOrientation, SetPosition,
 };
 use crate::geom::{self, Point2, Vector2};
