+++ src/draw/primitive/tri.rs
@@ -1,58 +1,60 @@
 use crate::draw::mesh::vertex::IntoPoint;
+use crate::draw::primitive::Primitive;
 use crate::draw::properties::spatial::{dimension, orientation, position};
 use crate::draw::properties::{
-    spatial, ColorScalar, Draw, Drawn, IntoDrawn, LinSrgba, Primitive, SetColor, SetDimensions,
+    spatial, ColorScalar, Draw, Drawn, IntoDrawn, LinSrgba, SetColor, SetDimensions,
     SetOrientation, SetPosition,
 };
 use crate::draw::{self, Drawing};
 use crate::geom::{self, Point3, Vector3};
 use crate::math::{BaseFloat, ElementWise};
-use std::{iter, slice};
+use std::ops;
 
-/// Properties related to drawing a **Quad**.
+/// Properties related to drawing a **Tri**.
 #[derive(Clone, Debug)]
-pub struct Quad<S = geom::scalar::Default> {
-    quad: geom::Quad<Point3<S>>,
+pub struct Tri<S = geom::scalar::Default> {
+    tri: geom::Tri<Point3<S>>,
     spatial: spatial::Properties<S>,
     color: Option<LinSrgba>,
 }
-// Quad-specific methods.
 
-impl<S> Quad<S> {
-    /// Use the given four points as the vertices (corners) of the quad.
-    pub fn points<P>(mut self, a: P, b: P, c: P, d: P) -> Self
+// Tri-specific methods.
+
+impl<S> Tri<S> {
+    /// Use the given three points as the vertices (corners) of the triangle.
+    pub fn points<P>(mut self, a: P, b: P, c: P) -> Self
     where
         P: IntoPoint<S>,
     {
         let a = a.into_point();
         let b = b.into_point();
         let c = c.into_point();
-        let d = d.into_point();
-        self.quad = geom::Quad([a, b, c, d]);
+        self.tri = geom::Tri([a, b, c]);
         self
     }
 }
 
 // Trait implementations.
 
-impl<S> IntoDrawn<S> for Quad<S>
+impl<S> IntoDrawn<S> for Tri<S>
 where
     S: BaseFloat,
 {
-    type Vertices = draw::mesh::vertex::IterFromPoints<geom::quad::Vertices<Point3<S>>, S>;
-    type Indices = iter::Cloned<slice::Iter<'static, usize>>;
+    type Vertices = draw::mesh::vertex::IterFromPoints<geom::tri::Vertices<Point3<S>>, S>;
+    type Indices = ops::Range<usize>;
     fn into_drawn(self, draw: Draw<S>) -> Drawn<S, Self::Vertices, Self::Indices> {
-        let Quad {
-            mut quad,
+        let Tri {
+            mut tri,
             spatial,
             color,
         } = self;
 
-        // If dimensions were specified, scale the points to those dimensions.
         let (maybe_x, maybe_y, maybe_z) = spatial.dimensions.to_scalars(&draw);
+
+        // If dimensions were specified, scale the points to those dimensions.
         if maybe_x.is_some() || maybe_y.is_some() || maybe_z.is_some() {
-            let cuboid = quad.bounding_cuboid();
-            let centroid = quad.centroid();
+            let cuboid = tri.bounding_cuboid();
+            let centroid = tri.centroid();
             let x_scale = maybe_x.map(|x| x / cuboid.w()).unwrap_or_else(S::one);
             let y_scale = maybe_y.map(|y| y / cuboid.h()).unwrap_or_else(S::one);
             let z_scale = maybe_z.map(|z| z / cuboid.d()).unwrap_or_else(S::one);
@@ -61,13 +63,12 @@ where
                 y: y_scale,
                 z: z_scale,
             };
-            let (a, b, c, d) = quad.into();
+            let (a, b, c) = tri.into();
             let translate = |v: Point3<S>| centroid + ((v - centroid).mul_element_wise(scale));
             let new_a = translate(a);
             let new_b = translate(b);
             let new_c = translate(c);
-            let new_d = translate(d);
-            quad = geom::Quad([new_a, new_b, new_c, new_d]);
+            tri = geom::Tri([new_a, new_b, new_c]);
         }
 
         // The color.
@@ -77,90 +78,82 @@ where
                     theme
                         .color
                         .primitive
-                        .get(&draw::theme::Primitive::Quad)
+                        .get(&draw::theme::Primitive::Tri)
                         .map(|&c| c.into_linear())
                 })
             })
             .unwrap_or(draw.theme(|t| t.color.default.into_linear()));
 
-        let points = quad.vertices();
+        let points = tri.vertices();
         let vertices = draw::mesh::vertex::IterFromPoints::new(points, color);
-        let indices = geom::quad::TRIANGLE_INDICES.iter().cloned();
+        let indices = 0..geom::tri::NUM_VERTICES as usize;
 
         (spatial, vertices, indices)
     }
 }
 
-impl<S> From<geom::Quad<Point3<S>>> for Quad<S>
+impl<S> From<geom::Tri<Point3<S>>> for Tri<S>
 where
     S: BaseFloat,
 {
-    fn from(quad: geom::Quad<Point3<S>>) -> Self {
+    fn from(tri: geom::Tri<Point3<S>>) -> Self {
         let spatial = <_>::default();
         let color = <_>::default();
-        Quad {
-            quad,
+        Tri {
+            tri,
             spatial,
             color,
         }
     }
 }
 
-impl<S> Default for Quad<S>
+impl<S> Default for Tri<S>
 where
     S: BaseFloat,
 {
     fn default() -> Self {
-        // Create a quad pointing towards 0.0 radians.
+        // Create a triangle pointing towards 0.0 radians.
         let zero = S::zero();
         let fifty = S::from(50.0).unwrap();
-        let left = -fifty;
-        let bottom = -fifty;
-        let right = fifty;
-        let top = fifty;
+        let thirty_three = S::from(33.0).unwrap();
         let a = Point3 {
-            x: left,
-            y: bottom,
+            x: -fifty,
+            y: thirty_three,
             z: zero,
         };
         let b = Point3 {
-            x: left,
-            y: top,
+            x: fifty,
+            y: zero,
             z: zero,
         };
         let c = Point3 {
-            x: right,
-            y: top,
+            x: -fifty,
+            y: -thirty_three,
             z: zero,
         };
-        let d = Point3 {
-            x: right,
-            y: bottom,
-            z: zero,
-        };
-        Quad::from(geom::Quad([a, b, c, d]))
+        Tri::from(geom::Tri([a, b, c]))
     }
 }
 
-impl<S> SetOrientation<S> for Quad<S> {
+impl<S> SetOrientation<S> for Tri<S> {
     fn properties(&mut self) -> &mut orientation::Properties<S> {
         SetOrientation::properties(&mut self.spatial)
     }
 }
 
-impl<S> SetPosition<S> for Quad<S> {
+impl<S> SetPosition<S> for Tri<S> {
     fn properties(&mut self) -> &mut position::Properties<S> {
         SetPosition::properties(&mut self.spatial)
     }
 }
 
-impl<S> SetDimensions<S> for Quad<S> {
+impl<S> SetDimensions<S> for Tri<S> {
     fn properties(&mut self) -> &mut dimension::Properties<S> {
         SetDimensions::properties(&mut self.spatial)
     }
 }
 
-impl<S> SetColor<ColorScalar> for Quad<S> {
+impl<S> SetColor<ColorScalar> for Tri<S> {
     fn rgba_mut(&mut self) -> &mut Option<LinSrgba> {
         SetColor::rgba_mut(&mut self.color)
     }
@@ -168,16 +161,16 @@ impl<S> SetColor<ColorScalar> for Quad<S> {
 
 // Primitive conversions.
 
-impl<S> From<Quad<S>> for Primitive<S> {
-    fn from(prim: Quad<S>) -> Self {
-        Primitive::Quad(prim)
+impl<S> From<Tri<S>> for Primitive<S> {
+    fn from(prim: Tri<S>) -> Self {
+        Primitive::Tri(prim)
     }
 }
 
-impl<S> Into<Option<Quad<S>>> for Primitive<S> {
-    fn into(self) -> Option<Quad<S>> {
+impl<S> Into<Option<Tri<S>>> for Primitive<S> {
+    fn into(self) -> Option<Tri<S>> {
         match self {
-            Primitive::Quad(prim) => Some(prim),
+            Primitive::Tri(prim) => Some(prim),
             _ => None,
         }
     }
@@ -185,15 +178,15 @@ impl<S> Into<Option<Quad<S>>> for Primitive<S> {
 
 // Drawing methods.
 
-impl<'a, S> Drawing<'a, Quad<S>, S>
+impl<'a, S> Drawing<'a, Tri<S>, S>
 where
     S: BaseFloat,
 {
-    /// Use the given points as the vertices (corners) of the quad.
-    pub fn points<P>(self, a: P, b: P, c: P, d: P) -> Self
+    /// Use the given points as the vertices (corners) of the triangle.
+    pub fn points<P>(self, a: P, b: P, c: P) -> Self
     where
         P: IntoPoint<S>,
     {
-        self.map_ty(|ty| ty.points(a, b, c, d))
+        self.map_ty(|ty| ty.points(a, b, c))
     }
 }
